Add Piece class to the TetrisPrototype model
Commit introduces Piece class defining pieces and their methods. In future the piece is about to be rotated, hence Piece#anchor is introduced (top-left point occupied by the piece).
Moreover, to be able to paint a piece on a canvas with an offset, Piece#translated_points method is here to do the (not so) heavy lifting.